导入Mysql

+++ md/stockdata.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-SYMBOL="001979.sz"
-OUTFILE=st.${SYMBOL}.data
+if [ -z "$1" ]; then
+    echo "usage: $0 <stock>"
+    exit -1
+fi
+
+SYMBOL=$1
+
+OUTFILE=st.${SYMBOL}.csv
 wget "http://table.finance.yahoo.com/table.csv?s=${SYMBOL}" -O $OUTFILE
 if [ ! -s $OUTFILE ]; then
     rm $OUTFILE
